+++ app/Console/Kernel.php
@@ -13,6 +13,7 @@ use ec5\Console\Commands\SystemStatsCommand;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
 use Illuminate\Support\Facades\App;
+use jdavidbakr\LaravelCacheGarbageCollector\LaravelCacheGarbageCollector;
 
 class Kernel extends ConsoleKernel
 {
@@ -28,7 +29,8 @@ class Kernel extends ConsoleKernel
         SystemProjectStorageCommand::class,
         SeedEntriesCommand::class,
         SeedMediaCommand::class,
-        SeedSuperadminCommand::class
+        SeedSuperadminCommand::class,
+        LaravelCacheGarbageCollector::class
     ];
 
     /**
@@ -40,13 +42,19 @@ class Kernel extends ConsoleKernel
             //grab system stats for the current day
             $schedule->command('system:stats')
                 ->dailyAt('01:00')
-                ->timezone('Europe/London')
+                ->timezone('UTC')
                 ->withoutOverlapping();
 
             //check storage available
             $schedule->command('system:check-storage')
                 ->dailyAt('07:00')
-                ->timezone('Europe/London')
+                ->timezone('UTC')
+                ->withoutOverlapping();
+
+            //clear laravel expired cache files on Sunday at 04:00 UTC
+            $schedule->command('cache:gc')
+                ->weeklyOn(0, '04:00')
+                ->timezone('UTC')
                 ->withoutOverlapping();
         } else {
             //run commands every hour locally (for debugging)