+++ routes/web.php
@@ -64,9 +64,6 @@ Route::group(['middleware' => 'auth.admin'], function () {
 
     Route::post('admin/update-user-project-role', 'Web\Admin\AdminUserRolesController@update');
 
-    Route::get('admin/tools/resize-entry-images', 'Web\Admin\Tools\ImageToolsController@resizeEntryImages');
-    Route::get('admin/tools/create-entry-extra-images', 'Web\Admin\Tools\ImageToolsController@createEntryExtraImages');
-
     Route::get('admin/tools/opcache', 'Web\Admin\Tools\PHPToolsController@resetOpcache');
     Route::get('admin/tools/projects-stats', 'Web\Admin\Tools\PHPToolsController@showProjectsStats');