+++ tests/Http/Controllers/Api/Entries/Delete/DeleteControllerEntriesS3Test.php
@@ -632,7 +632,7 @@ class DeleteControllerEntriesS3Test extends TestCase
                 $response[0]->assertExactJson([
                     "errors" => [[
                         "code" => "ec5_255",
-                        "title" => "Too many requests have been made. Please try again later.",
+                        "title" => "Too many requests. Try again later.",
                         "source" => "errors"
                     ]]
                 ]);